+++ frontend/src/components/icons/MoneyIcon.jsx
@@ -1,0 +1,24 @@
+/**
+ *
+ * @param {object} param
+ * @param {number} param.size
+ * @returns {JSX.Element}
+ */
+export function MoneyIcon({ size = 25 }) {
+  return (
+    <svg
+      width={`${size}`}
+      height={`${size}`}
+      viewBox="0 0 25 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clipPath="url(#clip0_219_109)">
+        <path
+          d="M15.1468 5.7C16.3927 4.6375 17.7083 3.04583 17.7083 1.04167V0H7.29163V1.04167C7.29163 3.04583 8.60725 4.63646 9.85308 5.7C4.95204 7.44688 1.04163 13.6708 1.04163 18.75C1.04163 22.1969 3.84475 25 7.29163 25H17.7083C21.1552 25 23.9583 22.1969 23.9583 18.75C23.9583 13.6708 20.0479 7.44688 15.1468 5.7ZM9.59058 2.08333H15.4062C14.8635 3.39167 13.4375 4.48958 12.5073 5.05417C11.7239 4.575 10.1625 3.47917 9.59163 2.08333H9.59058ZM17.7083 22.9167H7.29163C4.99371 22.9167 3.12496 21.0479 3.12496 18.75C3.12496 13.374 7.90621 7.30937 12.4979 7.30937C17.0895 7.30937 21.875 13.375 21.875 18.75C21.875 21.0479 20.0062 22.9167 17.7083 22.9167ZM10.4166 13.5417C10.4166 13.9354 10.6989 14.2687 11.0875 14.3333L14.2562 14.8615C15.6531 15.0938 16.6677 16.2917 16.6677 17.7083C16.6677 19.4312 15.2656 20.8333 13.5427 20.8333V21.875H11.4593V20.8333C9.73642 20.8333 8.33434 19.4312 8.33434 17.7083H10.4177C10.4177 18.2823 10.8854 18.75 11.4593 18.75H13.5427C14.1166 18.75 14.5843 18.2823 14.5843 17.7083C14.5843 17.3146 14.302 16.9812 13.9135 16.9167L10.7448 16.3885C9.34788 16.1562 8.33329 14.9583 8.33329 13.5417C8.33329 11.8187 9.73538 10.4167 11.4583 10.4167V9.375H13.5416V10.4167C15.2645 10.4167 16.6666 11.8187 16.6666 13.5417H14.5833C14.5833 12.9677 14.1156 12.5 13.5416 12.5H11.4583C10.8843 12.5 10.4166 12.9677 10.4166 13.5417Z"
+          fill="black"
+        />
+      </g>
+    </svg>
+  );
+}
