+++ frontend/src/components/icons/TimeIcon.jsx
@@ -1,0 +1,24 @@
+/**
+ *
+ * @param {object} param
+ * @param {number} param.size
+ * @returns {JSX.Element}
+ */
+export function TimeIcon({ size = 25 }) {
+  return (
+    <svg
+      width={`${size}`}
+      height={`${size}`}
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clipPath="url(#clip0_32_279)">
+        <path
+          d="M17.776 21.334C17.212 22.929 15.691 24 13.995 24H4.00503C2.83503 24 1.72603 23.49 0.961033 22.602C0.208033 21.727 -0.125967 20.577 0.044033 19.444C0.449033 16.746 1.92903 14.246 4.44503 12C1.92903 9.754 0.449033 7.254 0.044033 4.556C-0.125967 3.424 0.208033 2.273 0.961033 1.399C1.72603 0.51 2.83503 0 4.00503 0H13.995C15.691 0 17.212 1.071 17.776 2.666C17.961 3.187 17.688 3.758 17.168 3.942C16.647 4.129 16.076 3.854 15.892 3.334C15.609 2.536 14.847 2 13.996 2H4.00503C3.41803 2 2.86003 2.257 2.47703 2.703C2.09903 3.142 1.93803 3.694 2.02303 4.258C2.39803 6.757 3.94503 9.098 6.62103 11.216C6.86103 11.405 7.00103 11.695 7.00103 12C7.00103 12.305 6.86103 12.595 6.62103 12.784C3.94503 14.902 2.39803 17.242 2.02303 19.741C1.93803 20.305 2.09903 20.858 2.47703 21.297C2.86103 21.743 3.41803 22 4.00503 22H13.995C14.847 22 15.608 21.464 15.891 20.667C16.076 20.146 16.649 19.871 17.167 20.059C17.688 20.243 17.96 20.815 17.775 21.335L17.776 21.334ZM9.62603 13.944L10.134 14.349C11.306 15.281 12.861 16.78 13.608 18.624C13.733 18.933 13.696 19.283 13.51 19.559C13.324 19.834 13.014 20 12.681 20H5.31903C4.98703 20 4.67603 19.834 4.49003 19.559C4.30403 19.284 4.26703 18.934 4.39103 18.626C5.13703 16.773 6.69103 15.28 7.86303 14.353L8.38303 13.942C8.74703 13.653 9.26203 13.654 9.62603 13.944ZM8.99903 16.004C8.36803 16.513 7.62803 17.198 7.03403 18H10.962C10.368 17.202 9.63003 16.516 8.99903 16.004ZM24 12.001C24 15.31 21.309 18.001 18 18.001C14.691 18.001 12 15.31 12 12.001C12 8.692 14.691 6.001 18 6.001C21.309 6.001 24 8.692 24 12.001ZM22 12.001C22 9.795 20.206 8.001 18 8.001C15.794 8.001 14 9.795 14 12.001C14 14.207 15.794 16.001 18 16.001C20.206 16.001 22 14.207 22 12.001ZM19 11.587V10.001C19 9.449 18.553 9.001 18 9.001C17.447 9.001 17 9.449 17 10.001V12.001C17 12.266 17.105 12.521 17.293 12.708L18.293 13.708C18.488 13.903 18.744 14.001 19 14.001C19.256 14.001 19.512 13.903 19.707 13.708C20.098 13.317 20.098 12.685 19.707 12.294L19 11.587Z"
+          fill="black"
+        />
+      </g>
+    </svg>
+  );
+}
